Convert to proper python package
